Sync streams refactor - extract syncedStreamsLoop
So that i can re-use it with different data structures.

Right now if the rpc client craps out, we have no way to shut down the syncedStreams loop and restart it without booting up a whole new client. This opens up that possibility
